requirements.txt
selenium
webdriver-manager
python-dotenv
# slackweb